pyproject.toml
[tool.poetry]
name = "bnss"
version = "0.1.0"
description = "Music bot for Discord written in Python"
authors = ["atbuy <contact.atbuy@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8.1"
discord-py = "^2.3.1"
pydantic-settings = "^2.0.2"


[tool.poetry.group.dev.dependencies]
black = "^23.7.0"
isort = "^5.12.0"
flake8 = "^6.1.0"
pre-commit = "^3.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
